Commit for study only: whenever state update depends on previous state
If you use this approach, React will guarantee  that the state snapshot it gives you here  in this inner function,  will always be the latest state snapshot,  keeping all scheduled state updates in mind.  So this is the safer way  to ensure that you always operate  on the latest state snapshot.  So you should use this function syntax here  whenever your state update depends on the previous state.  That is something you should memorize.  If your state update depends on the previous state,  use this function form.